Fixed publication key error

+++ app/models/ppublication.rb
@@ -8,7 +8,7 @@ class Ppublication < ActiveRecord::Base
   def self.update(params)
     errors = ""
     self.where(:user_id => params[:id]).each do |pub|
-      if ((params[:data][pub.id.to_s]).key?("check"))
+      if params[:data].key?(pub.id.to_s) && params[:data][pub.id.to_s].key?("check")
         pub.destroy
       else
         pub.update_attributes((params[:data][pub.id.to_s])['data'])
